requirements.txt
Flask>=2.0.0
python-dotenv>=0.19.0
# Add other dependencies later, e.g., requests, pandas, scikit-learn, nltk, etc.